+++ packages/isar/lib/src/isar_link.dart
@@ -32,7 +32,7 @@ abstract class IsarLinkBase<OBJ> {
 abstract class IsarLink<OBJ> extends IsarLinkBase<OBJ> {
   /// Create an empty, unattached link. Make sure to provide the correct
   /// generic argument.
-  factory IsarLink() => isarCreateLink();
+  factory IsarLink() => IsarNative.newLink();
 
   /// The linked object or `null` if no object is linked.
   OBJ? get value;
@@ -46,7 +46,7 @@ abstract class IsarLink<OBJ> extends IsarLinkBase<OBJ> {
 abstract class IsarLinks<OBJ> extends IsarLinkBase<OBJ> implements Set<OBJ> {
   /// Create an empty, unattached link. Make sure to provide the correct
   /// generic argument.
-  factory IsarLinks() => isarCreateLinks();
+  factory IsarLinks() => IsarNative.newLinks();
 
   @override
   Future<void> load({bool overrideChanges = false});