+++ packages/isar/lib/src/web/isar_impl.dart
@@ -1,0 +1,101 @@
+import 'dart:async';
+
+import 'package:isar/isar.dart';
+
+import 'bindings.dart';
+
+const _zoneTxn = #zoneTxn;
+
+class IsarImpl extends Isar {
+  final IsarInstanceJs instance;
+  final List<Future> _activeAsyncTxns = [];
+
+  IsarImpl(String name, String schema, this.instance) : super(name, schema);
+
+  void requireNotInTxn() {
+    if (Zone.current[_zoneTxn] != null) {
+      throw IsarError(
+          'Cannot perform this operation from within an active transaction.');
+    }
+  }
+
+  void requireOpen() {
+    if (!isOpen) {
+      throw IsarError('Isar instance has already been closed');
+    }
+  }
+
+  Future<T> _txn<T>(
+      bool write, bool silent, Future<T> Function(Isar isar) callback) async {
+    requireOpen();
+    requireNotInTxn();
+
+    final completer = Completer();
+    _activeAsyncTxns.add(completer.future);
+
+    final txn = instance.beginTxn(write);
+
+    final zone = Zone.current.fork(
+      zoneValues: {_zoneTxn: txn},
+    );
+
+    T result;
+    try {
+      result = await zone.run(() => callback(this));
+      await txn.commit().wait();
+    } catch (e) {
+      txn.abort();
+      rethrow;
+    } finally {
+      completer.complete();
+      _activeAsyncTxns.remove(completer.future);
+    }
+
+    return result;
+  }
+
+  @override
+  Future<T> txn<T>(Future<T> Function(Isar isar) callback) {
+    return _txn(false, false, callback);
+  }
+
+  @override
+  Future<T> writeTxn<T>(Future<T> Function(Isar isar) callback,
+      {bool silent = false}) {
+    return _txn(true, silent, callback);
+  }
+
+  @override
+  T txnSync<T>(T Function(Isar isar) callback) => throw UnimplementedError();
+
+  @override
+  T writeTxnSync<T>(T Function(Isar isar) callback, {bool silent = false}) =>
+      throw UnimplementedError();
+
+  Future<T> getTxn<T>(bool write, Future<T> Function(IsarTxnJs txn) callback) {
+    IsarTxnJs? currentTxn = Zone.current[_zoneTxn];
+    if (currentTxn != null) {
+      if (write && !currentTxn.write) {
+        throw IsarError(
+            'Operation cannot be performed within a read transaction.');
+      }
+      return callback(currentTxn);
+    } else if (!write) {
+      return _txn(false, false, (isar) {
+        return callback(Zone.current[_zoneTxn]!);
+      });
+    } else {
+      throw IsarError('Write operations require an explicit transaction.');
+    }
+  }
+
+  @override
+  Future<bool> close({bool deleteFromDisk = false}) async {
+    requireOpen();
+    requireNotInTxn();
+    await Future.wait(_activeAsyncTxns);
+    await super.close();
+    await instance.close(deleteFromDisk).wait();
+    return true;
+  }
+}