+++ packages/isar/lib/src/web/isar_web.dart
@@ -1,9 +1,16 @@
-import 'dart:js';
+import 'dart:js_util';
 import 'dart:typed_data';
 
 import 'package:isar/isar.dart';
-import 'package:isar/src/isar_native_def.dart';
+import 'package:isar/src/isar_native_interface.dart';
+import 'package:isar/src/web/js_converter.dart';
 import 'package:meta/meta.dart';
+
+import 'open.dart';
+
+const isarMinId = -9007199254740991;
+const isarMaxId = 9007199254740991;
+const isarAutoIncrementId = double.negativeInfinity;
 
 /// @nodoc
 @protected
@@ -19,53 +26,81 @@ typedef IsarBinaryReader = dynamic;
 
 /// @nodoc
 @protected
-typedef IsarBinaryWriter = dynamic;
+// ignore: constant_identifier_names
+const dynamic IsarBinaryWriter = null;
 
-/// @nodoc
-@protected
-typedef IsarJsObject = JsObject;
+typedef IsarJsConverter = JsConverter;
 
-Uint8List _isarBufAsBytes(IsarBytePointer pointer, int length) {
-  throw UnimplementedError();
+class _IsarWeb implements IsarNativeInterface {
+  const _IsarWeb();
+
+  @override
+  Uint8List bufAsBytes(IsarBytePointer pointer, int length) {
+    throw UnimplementedError();
+  }
+
+  @override
+  void initializeLibraries({Map<String, String> libraries = const {}}) {
+    throw UnimplementedError();
+  }
+
+  @override
+  dynamic jsObjectGet(Object o, Object key) {
+    getProperty(o, key);
+  }
+
+  @override
+  void jsObjectSet(Object o, Object key, dynamic value) {
+    setProperty(o, key, value);
+  }
+
+  @override
+  dynamic newJsObject() {
+    return newObject();
+  }
+
+  @override
+  IsarLink<OBJ> newLink<OBJ>() {
+    throw UnimplementedError();
+  }
+
+  @override
+  IsarLinks<OBJ> newLinks<OBJ>() {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<Isar> open({
+    required String directory,
+    required String name,
+    required bool relaxedDurability,
+    required List<CollectionSchema> schemas,
+  }) {
+    return openIsar(
+      name: name,
+      directory: directory,
+      relaxedDurability: relaxedDurability,
+      schemas: schemas,
+    );
+  }
+
+  @override
+  Isar openSync({
+    required String directory,
+    required String name,
+    required bool relaxedDurability,
+    required List<CollectionSchema> schemas,
+  }) {
+    throw UnimplementedError();
+  }
+
+  @override
+  List<String> splitWords(String value) {
+    throw UnimplementedError();
+  }
 }
 
 /// @nodoc
 @protected
-const IsarBufAsBytes isarBufAsBytes = _isarBufAsBytes;
-
-/// @nodoc
-//@protected
-//const IsarSplitWords isarSplitWords = splitWords;
-
-/// @nodoc
-//@protected
-//const IsarOpen isarOpen = openIsar;
-
-Isar _isarOpenSync({
-  required String directory,
-  required String name,
-  required bool relaxedDurability,
-  required List<CollectionSchema> schemas,
-}) {
-  throw UnimplementedError();
-}
-
-/// @nodoc
-//@protected
-const IsarOpenSync isarOpenSync = _isarOpenSync;
-
-/// @nodoc
-//@protected
-//const IsarCreateLink isarCreateLink = createIsarLink;
-
-/// @nodoc
-//@protected
-//const IsarCreateLinks isarCreateLinks = createIsarLinks;
-
-IsarJsObject _isarCreateJsObject() {
-  return JsObject(context['Object']);
-}
-
-/// @nodoc
-@protected
-const IsarCreateJsObject isarCreateJsObject = _isarCreateJsObject;
+// ignore: constant_identifier_names
+const IsarNative = _IsarWeb();