+++ packages/isar_generator/lib/src/code_gen/query_sort_by_generator.dart
@@ -12,11 +12,11 @@ String generateSortBy(ObjectInfo oi) {
 
     code += '''
     QueryBuilder<${oi.dartName}, ${oi.dartName}, QAfterSortBy>sortBy${property.dartName.capitalize()}() {
-      return addSortByInternal('${property.dartName.esc}', Sort.asc);
+      return addSortByInternal('${property.isarName.esc}', Sort.asc);
     }
     
     QueryBuilder<${oi.dartName}, ${oi.dartName}, QAfterSortBy>sortBy${property.dartName.capitalize()}Desc() {
-      return addSortByInternal('${property.dartName.esc}', Sort.desc);
+      return addSortByInternal('${property.isarName.esc}', Sort.desc);
     }''';
   }
 
@@ -30,11 +30,11 @@ String generateSortBy(ObjectInfo oi) {
 
     code += '''
     QueryBuilder<${oi.dartName}, ${oi.dartName}, QAfterSortBy>thenBy${property.dartName.capitalize()}() {
-      return addSortByInternal('${property.dartName.esc}', Sort.asc);
+      return addSortByInternal('${property.isarName.esc}', Sort.asc);
     }
     
     QueryBuilder<${oi.dartName}, ${oi.dartName}, QAfterSortBy>thenBy${property.dartName.capitalize()}Desc() {
-      return addSortByInternal('${property.dartName.esc}', Sort.desc);
+      return addSortByInternal('${property.isarName.esc}', Sort.desc);
     }''';
   }
   code += '}';